added spring security and jwt

+++ src/main/java/com/benjamin/Banking_app/Dto/TransferRequest.java
@@ -8,5 +8,6 @@ import lombok.Setter;
 public class TransferRequest {
     private Long fromAccountId;
     private Long toAccountId;
-    private Long amount;
+    private Double amount;
+
 }
